rename and position
